moved actors_list package to ui presentation/package
